追加のアクションをテーブルのヘッダー内に移動した
